rename: profile container 파일 위치 변경
