--- src/pages/ProfilePage/Containers/EditProfile/styles.js
@@ -1,18 +1,0 @@
-import styled from '@emotion/styled/macro';
-
-const PasswordContainer = styled.form`
-  display: flex;
-  width: 60%;
-  flex-direction: column;
-  justify-content: space-between;
-  padding: 0 0 7rem 0;
-`;
-
-const NicknameContainer = styled.form`
-  display: flex;
-  width: 60%;
-  flex-direction: column;
-  padding: 3rem 0;
-`;
-
-export { PasswordContainer, NicknameContainer };
