+++ src/pages/ProfilePage/Containers/NicknameEdit/styles.js
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled/macro';
+
+const NicknameContainer = styled.form`
+  display: flex;
+  width: 60%;
+  flex-direction: column;
+  padding: 3rem 0;
+`;
+
+export { NicknameContainer };
