+++ src/pages/ProfilePage/Containers/PasswordEdit/styles.js
@@ -8,11 +8,4 @@ const PasswordContainer = styled.form`
   padding: 0 0 7rem 0;
 `;
 
-const NicknameContainer = styled.form`
-  display: flex;
-  width: 60%;
-  flex-direction: column;
-  padding: 3rem 0;
-`;
-
-export { PasswordContainer, NicknameContainer };
+export { PasswordContainer };
